debug: remove <Plug> mappings on stop
Remove <Plug> mapping on stop. This was mistakenly removed in a recent
comment that changed the code that removes mapping so that only the key
mappings were removed.

Add a test to ensure <Plug> mappings are removed.

It turns out that the previous commit should have added the removal of
key mappings and kept the removal of the <Plug> mappings.

This also fixes a bug with the previous implementation of removal of
<Plug> mappings to ensure that the first <Plug> mapping is also removed.